apps/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.1
redis==5.0.1
pyswisseph==2.10.3.2
pinecone-client==2.2.4
sqlalchemy==2.0.19
asyncpg==0.28.0
alembic==1.11.1
passlib==1.7.4
python-jose==3.3.0
# Assumption: Versions based on compatibility; doc silent on versions.